Moving backtrackers into their own folder. Making them subclasses of a common backtracker
