+++ src/pu/backtrackers/backtrack_edits_unit.cc
@@ -1,0 +1,124 @@
+/*
+ * Copyright (C) 2015 BIO-DIKU.
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; either version 2 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc., 51
+ * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
+ *
+ * http://www.gnu.org/copyleft/gpl.html
+ */
+
+#include "backtrack_edits_unit.h"
+
+#include <utility>
+#include <iostream>
+
+using namespace std;
+
+
+BacktrackEditsUnit::BacktrackEditsUnit(
+    const Modifiers &modifiers,
+    const std::string& pattern
+) :
+    BacktrackUnit(modifiers, pattern)
+{ }
+
+bool BacktrackEditsUnit::FindMatch() {
+  if (sequence_iterator_ == sequence_iterator_end_) return false;
+
+  if (stay_at_pos_) {
+    // TODO(Ras) This is a little ugly. Clean up when not tired
+    if (last_found_index_ != 0 && last_found_index_ + 1 >= last_found_matches_.size()) {
+      return false;
+    }
+
+    if (last_found_matches_.empty()) {
+      int E = modifiers_.max_edits_;
+
+      CollectMatches(sequence_iterator_, pattern_.cbegin(), E, 0,0,0);
+      if (last_found_matches_.empty()) {
+        last_found_index_++;
+        return false;
+      } else {
+        return true;
+      }
+    }
+    last_found_index_++;
+    return last_found_index_ < last_found_matches_.size();
+  }
+
+  if (last_found_index_ + 1 >= last_found_matches_.size()) {
+
+    // All previously identified matches have been iterated through. Time to
+    // look for new ones.
+    last_found_matches_.clear();
+    last_found_index_ = 0;
+
+    int E = modifiers_.max_edits_;
+
+    while (sequence_iterator_ != sequence_iterator_end_) {
+      CollectMatches(sequence_iterator_, pattern_.cbegin(), E, 0,0,0);
+
+      ++sequence_iterator_;
+
+      if (!last_found_matches_.empty())
+        return true;
+    }
+
+    return false;
+
+  } else {
+    last_found_index_++;
+    return true;
+  }
+}
+
+void BacktrackEditsUnit::CollectMatches(
+    std::string::const_iterator seq_it,
+    std::string::const_iterator pat_it,
+    const int E_left,
+    const int M_used, const int I_used, const int D_used
+) {
+
+  if (pat_it == pattern_.cend()) {
+    last_found_matches_.insert(
+        Match(sequence_iterator_, pattern_.length()+I_used-D_used, M_used+I_used+D_used));
+    return;
+  }
+
+  if (seq_it == sequence_iterator_end_)
+    return;
+
+  if (modifiers_.res_matcher_.Match(*seq_it,*pat_it)) {
+    CollectMatches(seq_it + 1, pat_it + 1, E_left, M_used, I_used, D_used);
+    return;
+  }
+
+  if (E_left > 0) {
+    CollectMatches(seq_it + 1, pat_it + 1, E_left - 1, M_used+1, I_used  , D_used  ); // Check if mismatch
+    CollectMatches(seq_it    , pat_it + 1, E_left - 1, M_used  , I_used  , D_used+1); // Check if deletion
+    CollectMatches(seq_it + 1, pat_it    , E_left - 1, M_used  , I_used+1, D_used  ); // Check if insertion
+  }
+
+
+}
+
+std::ostream& operator<<(std::ostream& os, const BacktrackEditsUnit& obj) {
+  return obj.Print(os);
+}
+
+std::unique_ptr<PatternUnit> BacktrackEditsUnit::Clone() const {
+  std::unique_ptr<PatternUnit> ret( new BacktrackEditsUnit(modifiers_, pattern_) );
+  return std::move(ret);
+}
+
