+++ src/pu/backtrackers/backtrack_edits_unit.h
@@ -26,8 +26,8 @@
 #include <string>
 #include <iostream>
 
-#include "pattern_unit.h"
-#include "../match.h"
+#include "match.h"
+#include "backtrack_unit.h"
 
 /**
  * A pattern unit that matches a specified string pattern using a backtracking algorithm. The
@@ -35,45 +35,14 @@
  * The algorithm is conceptually similar to the one found in scan_for_matches, but bugs have been
  * corrected.
  */
-class BacktrackEditsUnit: public PatternUnit{
+class BacktrackEditsUnit: public BacktrackUnit{
  public:
   /** Construct a BacktrackEditsUnit with the specified modifiers and pattern-string. */
   BacktrackEditsUnit(const Modifiers &modifiers, const std::string& pattern);
 
-  void Initialize(
-      std::string::const_iterator pos,
-      std::string::const_iterator max_pos,
-      bool stay_at_pos = false) override;
-
   virtual bool FindMatch() override;
 
-  const Match& GetMatch() const override;
-
-  std::ostream& Print(std::ostream &os) const override;
-
  protected:
-  /** The pattern-string to search for. */
-  const std::string           pattern_;
-
-  /** The start sequence iterator. */
-  std::string::const_iterator sequence_iterator_;
-
-  /** The end sequence iterator. */
-  std::string::const_iterator sequence_iterator_end_;
-
-  /** Indicates whether FindMatch should search for matches starting at other positions than the
-   * one indicated in Initialize */
-  bool stay_at_pos_;
-
-  /** The backtracking collects a set of matches for each starting-position in the sequence. Its
-   * essential that this is a set (no duplicates) as backtracking can reach the same match via
-   * different paths, but only one should be reported by FindMatch. Furthermore, its convenient
-   * that the ordering of matches is well-defined (first, sorted by start-position, second by
-   * end-position). */
-  std::set<Match> last_found_matches_;
-
-  /** The index of the match in last_found_matches_ that GetMatch should return. */
-  size_t last_found_index_;
 
   /** A recursive function that fills up last_found_matches_. This is the core of the
    * backtracking algorithm. */
