+++ src/pu/backtrackers/backtrack_unit.h
@@ -26,8 +26,8 @@
 #include <string>
 #include <iostream>
 
-#include "pattern_unit.h"
-#include "../match.h"
+#include "pu/pattern_unit.h"
+#include "match.h"
 
 /**
  * A pattern unit that matches a specified string pattern using a backtracking algorithm. The
@@ -45,11 +45,11 @@ class BacktrackUnit: public PatternUnit{
       std::string::const_iterator max_pos,
       bool stay_at_pos = false) override;
 
-  virtual bool FindMatch() override;
+  virtual bool FindMatch() override = 0;
 
   const Match& GetMatch() const override;
 
-  std::ostream& Print(std::ostream &os) const override;
+  virtual std::ostream& Print(std::ostream &os) const override = 0;
 
  protected:
   /** The pattern-string to search for. */
@@ -75,15 +75,8 @@ class BacktrackUnit: public PatternUnit{
   /** The index of the match in last_found_matches_ that GetMatch should return. */
   size_t last_found_index_;
 
-  /** A recursive function that fills up last_found_matches_. This is the core of the
-   * backtracking algorithm. */
-  void CollectMatches(std::string::const_iterator seq_it,
-                      std::string::const_iterator pat_it,
-                      const int M_left, const int I_left, const int D_left,
-                      const int M_used, const int I_used, const int D_used);
-
 protected:
-  virtual std::unique_ptr<PatternUnit> Clone() const override;
+  virtual std::unique_ptr<PatternUnit> Clone() const override = 0;
 };
 
 std::ostream& operator<<(std::ostream& os, const BacktrackUnit& obj);
